src/requirements.txt
Flask
flask-cors
python-dotenv
langchain
langchain-community
langchain-huggingface
manim
faiss-cpu
huggingface-hub
torch
